tech(quest-detail): first pass at revamp
