+++ package.json
@@ -28,9 +28,10 @@
   "types": "dist/index.d.ts",
   "devDependencies": {
     "@chromatic-com/storybook": "^1.9.0",
-    "@mantine/core": "^7.5.1",
     "@hyperplay/chains": "^0.3.0",
-    "@hyperplay/ui": "^1.8.13-alpha.0",
+    "@hyperplay/ui": "file:/Users/eliobriceno/Documents/hyperplay/ui-library/hyperplay-ui-1.8.17-alpha.0.tgz",
+    "@hyperplay/utils": "^0.2.7",
+    "@mantine/core": "^7.5.1",
     "@storybook/addon-essentials": "^8.3.4",
     "@storybook/addon-interactions": "^8.3.4",
     "@storybook/addon-links": "^8.3.4",
@@ -39,7 +40,6 @@
     "@storybook/react": "^8.3.4",
     "@storybook/react-vite": "^8.3.4",
     "@storybook/test": "^8.3.4",
-    "@hyperplay/utils": "^0.2.7",
     "@tanstack/query-core": "^5.59.13",
     "@tanstack/react-query": "^5.51.23",
     "@types/jest": "^29.5.14",
@@ -54,11 +54,11 @@
     "eslint-plugin-react": "^7.35.0",
     "eslint-plugin-storybook": "^0.9.0",
     "i18next": "^23.12.3",
+    "jest": "^29.7.0",
+    "mobx": "^6.13.3",
     "postcss": "^8.4.47",
     "postcss-preset-mantine": "^1.17.0",
     "postcss-simple-vars": "^7.0.1",
-    "jest": "^29.7.0",
-    "mobx": "^6.13.3",
     "prettier": "^3.3.3",
     "prop-types": "^15.8.1",
     "react": "^18.3.1",
