+++ src/components/PlayStreakEligibilityWrapper/index.module.scss
@@ -1,0 +1,5 @@
+.syncSuccess {
+  display: flex;
+  align-items: center;
+  gap: var(--space-sm);
+}
